requirements.txt
certifi==2021.10.8
click==8.0.4
dnspython==2.2.1
Flask==2.0.3
Flask-PyMongo==2.3.0
itsdangerous==2.1.1
Jinja2==3.0.3
MarkupSafe==2.1.1
pymongo==4.1.0
python-dotenv==0.20.0
Werkzeug==2.0.3
